Extract Windowed Serde tests into public package
